backend/orchestrator/requirements.txt
flask
requests
flask-cors
psycopg2-binary